Restructure audit package into dedicated submodules
Move core components into specialized directories (queue, report, monitor, etc.) and remove redundant files. Update test imports to reflect new structure. This improves code organization, enhances modularity, and simplifies maintenance by grouping related functionality. The package.json in audit-db now includes explicit test configuration scripts for CI and watch modes.
