+++ packages/audit/src/error/database-error-logger.ts
