+++ packages/audit/src/error/error-handling.ts
