+++ packages/audit/src/index.ts
@@ -1,20 +1,20 @@
 export * from './audit.js'
 export * from './types.js'
 export * from './crypto.js'
-export * from './gdpr-compliance.js'
-export * from './gdpr-utils.js'
+export * from './gdpr/gdpr-compliance.js'
+export * from './gdpr/gdpr-utils.js'
 export * from './validation.js'
-export * from './event-types.js'
-export * from './event-categorization.js'
+export * from './event/event-types.js'
+export * from './event/event-categorization.js'
 export * from './retry.js'
-export * from './circuit-breaker.js'
-export * from './dead-letter-queue.js'
-export * from './reliable-processor.js'
+export * from './queue/circuit-breaker.js'
+export * from './queue/dead-letter-queue.js'
+export * from './queue/reliable-processor.js'
 
 // Export monitoring with explicit exports to avoid conflicts
-export { MonitoringService, ConsoleAlertHandler } from './monitoring.js'
-export type { AlertHandler } from './monitoring.js'
-export type { Alert, AlertSeverity, AlertType } from './monitoring-types.js'
+export { MonitoringService, ConsoleAlertHandler } from './monitor/monitoring.js'
+export type { AlertHandler } from './monitor/monitoring.js'
+export type { Alert, AlertSeverity, AlertType } from './monitor/monitoring-types.js'
 
 // Export health check with explicit exports to avoid conflicts
 export {
@@ -24,13 +24,13 @@ export {
 	QueueHealthCheck,
 	ProcessingHealthCheck,
 	CircuitBreakerHealthCheck,
-} from './health-check.js'
+} from './monitor/health-check.js'
 
 // Export compliance reporting services
-export * from './compliance-reporting.js'
-export * from './data-export.js'
-export * from './scheduled-reporting.js'
+export * from './report/compliance-reporting.js'
+export * from './report/data-export.js'
+export * from './report/scheduled-reporting.js'
 
 // Export error handling and logging services
-export * from './error-handling.js'
-export { DatabaseErrorLogger } from './database-error-logger.js'
+export * from './error/error-handling.js'
+export { DatabaseErrorLogger } from './error/database-error-logger.js'
